+++ kernel/src/checkpoint/tests.rs
@@ -1,13 +1,120 @@
-use crate::{
-    actions::{Add, Metadata, Protocol, Remove},
-    engine::default::{executor::tokio::TokioBackgroundExecutor, DefaultEngine},
-    utils::test_utils::Action,
-    DeltaResult, Table,
+use std::{
+    path::PathBuf,
+    sync::{atomic::Ordering, Arc},
+    time::Duration,
 };
+
+use crate::actions::{Add, Metadata, Protocol, Remove};
+use crate::arrow::array::{ArrayRef, StructArray};
+use crate::arrow::datatypes::{DataType, Schema};
+use crate::checkpoint::{deleted_file_retention_timestamp_with_time, CheckpointWriter};
+use crate::engine::arrow_data::ArrowEngineData;
+use crate::engine::{
+    default::{executor::tokio::TokioBackgroundExecutor, DefaultEngine},
+    sync::SyncEngine,
+};
+use crate::snapshot::Snapshot;
+use crate::utils::test_utils::Action;
+use crate::DeltaResult;
+use crate::Engine;
+use crate::Table;
+
+use arrow_53::{
+    array::{create_array, RecordBatch},
+    datatypes::Field,
+};
+
 use object_store::{memory::InMemory, path::Path, ObjectStore};
-use std::sync::{atomic::Ordering, Arc};
 use test_utils::delta_path_for_version;
 use url::Url;
+
+#[test]
+fn test_deleted_file_retention_timestamp() -> DeltaResult<()> {
+    let now = Duration::from_secs(1000).as_millis() as i64;
+
+    // Test cases
+    let test_cases = [
+        // Default case (7 days)
+        (None, now - (7 * 24 * 60 * 60 * 1000)),
+        // Zero retention
+        (Some(Duration::from_secs(0)), now),
+        // Custom retention (2000 seconds)
+        // This results in a negative timestamp which is valid - as it just means that
+        // the retention window extends to before UNIX epoch.
+        (Some(Duration::from_secs(2000)), now - (2000 * 1000)),
+    ];
+
+    for (retention, expected) in test_cases {
+        let result =
+            deleted_file_retention_timestamp_with_time(retention, Duration::from_secs(1000))?;
+        assert_eq!(result, expected);
+    }
+
+    Ok(())
+}
+
+fn create_test_snapshot(engine: &dyn Engine) -> DeltaResult<Arc<Snapshot>> {
+    let path = std::fs::canonicalize(PathBuf::from("./tests/data/app-txn-no-checkpoint/"));
+    let url = url::Url::from_directory_path(path.unwrap()).unwrap();
+    let table = Table::new(url);
+    Ok(Arc::new(table.snapshot(engine, None)?))
+}
+
+#[test]
+fn test_create_checkpoint_metadata_batch_when_v2_checkpoints_is_supported() -> DeltaResult<()> {
+    let engine = SyncEngine::new();
+    let version = 10;
+    let writer = CheckpointWriter::new(create_test_snapshot(&engine)?);
+
+    // Test with is_v2_checkpoint = true
+    let result = writer.create_checkpoint_metadata_batch(version, &engine, true)?;
+    assert!(result.is_some());
+    let checkpoint_data = result.unwrap()?;
+
+    // Check selection vector has one true value
+    assert_eq!(checkpoint_data.selection_vector, vec![true]);
+
+    // Verify the underlying EngineData contains the expected CheckpointMetadata action
+    let arrow_engine_data = ArrowEngineData::try_from_engine_data(checkpoint_data.data)?;
+    let record_batch = arrow_engine_data.record_batch();
+
+    // Build the expected RecordBatch
+    // Note: The schema is a struct with a single field "checkpointMetadata" of type struct
+    // containing a single field "version" of type long
+    let expected_schema = Arc::new(Schema::new(vec![Field::new(
+        "checkpointMetadata",
+        DataType::Struct(vec![Field::new("version", DataType::Int64, false)].into()),
+        false,
+    )]));
+    let expected = RecordBatch::try_new(
+        expected_schema,
+        vec![Arc::new(StructArray::from(vec![(
+            Arc::new(Field::new("version", DataType::Int64, false)),
+            create_array!(Int64, [version]) as ArrayRef,
+        )]))],
+    )
+    .unwrap();
+
+    assert_eq!(*record_batch, expected);
+    assert_eq!(writer.actions_count.load(Ordering::Relaxed), 1);
+
+    Ok(())
+}
+
+#[test]
+fn test_create_checkpoint_metadata_batch_when_v2_checkpoints_not_supported() -> DeltaResult<()> {
+    let engine = SyncEngine::new();
+    let writer = CheckpointWriter::new(create_test_snapshot(&engine)?);
+
+    // Test with is_v2_checkpoint = false
+    let result = writer.create_checkpoint_metadata_batch(10, &engine, false)?;
+
+    // No checkpoint metadata action should be created for V1 checkpoints
+    assert!(result.is_none());
+    assert_eq!(writer.actions_count.load(Ordering::Relaxed), 0);
+
+    Ok(())
+}
 
 /// TODO(#855): Merge copies and move to `test_utils`
 /// Create an in-memory store and return the store and the URL for the store's _delta_log directory.
@@ -154,8 +261,8 @@ fn test_v1_checkpoint_latest_version_by_default() -> DeltaResult<()> {
     // contain any true values, as the add action is removed in a following commit.
     assert!(data_iter.next().is_none());
 
-    assert_eq!(writer.total_actions_counter.load(Ordering::Relaxed), 4);
-    assert_eq!(writer.add_actions_counter.load(Ordering::Relaxed), 1);
+    assert_eq!(writer.actions_count.load(Ordering::Relaxed), 4);
+    assert_eq!(writer.add_actions_count.load(Ordering::Relaxed), 1);
 
     // TODO(#850): Finalize and verify _last_checkpoint
     Ok(())
@@ -207,8 +314,8 @@ fn test_v1_checkpoint_specific_version() -> DeltaResult<()> {
     // No more data should exist because we only requested version 0
     assert!(data_iter.next().is_none());
 
-    assert_eq!(writer.total_actions_counter.load(Ordering::Relaxed), 2);
-    assert_eq!(writer.add_actions_counter.load(Ordering::Relaxed), 0);
+    assert_eq!(writer.actions_count.load(Ordering::Relaxed), 2);
+    assert_eq!(writer.add_actions_count.load(Ordering::Relaxed), 0);
 
     // TODO(#850): Finalize and verify _last_checkpoint
     Ok(())
@@ -270,8 +377,8 @@ fn test_v2_checkpoint_supported_table() -> DeltaResult<()> {
     // No more data should exist
     assert!(data_iter.next().is_none());
 
-    assert_eq!(writer.total_actions_counter.load(Ordering::Relaxed), 5);
-    assert_eq!(writer.add_actions_counter.load(Ordering::Relaxed), 1);
+    assert_eq!(writer.actions_count.load(Ordering::Relaxed), 5);
+    assert_eq!(writer.add_actions_count.load(Ordering::Relaxed), 1);
 
     // TODO(#850): Finalize and verify _last_checkpoint
     Ok(())
